push-scroller: update to Epiphany master
Update to Epiphany master and also fix the ugly "jump when drag starts" bug.

Bug #605702

+++ extensions/push-scroller/ephy-push-scroller.h
@@ -62,8 +62,8 @@ EphyPushScroller       *ephy_push_scroller_new		 (EphyWindow *window);
 
 void			ephy_push_scroller_start	 (EphyPushScroller *scroller,
 							  EphyEmbed *embed,
-							  int x,
-							  int y);
+							  gdouble x,
+							  gdouble y);
 
 void			ephy_push_scroller_stop		 (EphyPushScroller *scroller,
 							  guint32 timestamp);
